+++ CHANGELOG.md
@@ -12,6 +12,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 - Documentation for OMICS variants and updating a case
 - Include both creation and last modification dates in gene panels pages
+- Moved code to collect MT copy number stats for the MT report to the chanjo extension
 ### Fixed
 - Broken heading anchors in the documentation (`admin-guide/login-system.md` and `admin-guide/setup-scout.md` files)
 - Avoid open login redirect attacks by always redirecting to cases page upon user login